// src/data/repos.generated.ts
export type Repo = {
  name: string;
  description: string | null;
  language: string | null;
  updatedAt: string;
  isPrivate: boolean;
  stars: number;
  url: string | null;
  homepage: string | null;
};